ai-knowledge-digest/requirements.txt
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
requests==2.31.0
beautifulsoup4==4.12.2
feedparser==6.0.10
youtube-transcript-api==0.6.2
openai==1.3.5
schedule==1.2.0
python-dateutil==2.8.2
email-validator==2.1.0
jinja2==3.1.2
markdown==3.5.1
pytest==7.4.3
black==23.12.1
flake8==6.1.0